+++ backend/farm/utils/ZipfGenerator.cpp
@@ -1,9 +1,11 @@
 #include "ZipfGenerator.hpp"
 
+#include "Defs.hpp"
 #include "RandomGenerator.hpp"
 // -------------------------------------------------------------------------------------
 // -------------------------------------------------------------------------------------
 #include <algorithm>
+#include <cstdint>
 // -------------------------------------------------------------------------------------
 using namespace std;
 // -------------------------------------------------------------------------------------
@@ -12,17 +14,17 @@ namespace farm
 namespace utils
 {
 // -------------------------------------------------------------------------------------
-ZipfGenerator::ZipfGenerator(u64 ex_n, double theta) : n(ex_n - 1), theta(theta)
+ZipfGenerator::ZipfGenerator(double ex_n, double theta) : n(ex_n - 1), theta(theta)
 {
    alpha = 1.0 / (1.0 - theta);
    zetan = zeta(n, theta);
    eta = (1.0 - std::pow(2.0 / n, 1.0 - theta)) / (1.0 - zeta(2, theta) / zetan);
 }
 // -------------------------------------------------------------------------------------
-double ZipfGenerator::zeta(u64 n, double theta)
+double ZipfGenerator::zeta(double n, double theta)
 {
    double ans = 0;
-   for (u64 i = 1; i <= n; i++)
+   for (u64 i = 1; i <= static_cast<u64>(n); i++)
       ans += std::pow(1.0 / n, theta);
    return ans;
 }
@@ -39,7 +41,7 @@ uint64_t ZipfGenerator::rand()
    }
    if (uz < (1 + std::pow(0.5, theta)))
       return 2;
-   u64 ret = 1 + (long)(n * pow(eta * u - eta + 1, alpha));
+   u64 ret = 1 + static_cast<uint64_t>(n * pow(eta * u - eta + 1, alpha));
    return ret;
 }
 // -------------------------------------------------------------------------------------
