+++ backend/farm/utils/ZipfGenerator.hpp
@@ -17,16 +17,16 @@ class ZipfGenerator
 {
    // -------------------------------------------------------------------------------------
   private:
-   u64 n;
+   double n;
    double theta;
    // -------------------------------------------------------------------------------------
    double alpha, zetan, eta;
    // -------------------------------------------------------------------------------------
-   double zeta(u64 n, double theta);
+   double zeta(double n, double theta);
 
   public:
    // [0, n)
-   ZipfGenerator(uint64_t ex_n, double theta);
+   ZipfGenerator(double ex_n, double theta);
    // uint64_t rand(u64 new_n);
    uint64_t rand();
 };
